Add release notes (NOT FINAL)
Added a non-final copy of the release notes, as well as another copy of the changelog. They are separate documents because the release notes are only relevant for this release, while  the changelog will be updated as we go. If we like, we can also keep copies of the various release notes for posterity's sake.